restructure repo with task file and all app code inside package for easier testing
