+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from starlette.testclient import TestClient
 
-from app import app
+from icehockey_rules.app import app
 
 
 @pytest.fixture(scope="function")
